# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CTWall Backend API
  version: 1.0.0
  description: |
    API for ChainThreatWall (CTWall) - SBOM Management and Security Analysis System.
    Adheres to OWASP ASVS Level 2 Security Requirements.
    Uses RFC 7807 for error handling.
servers:
  - url: https://api.ctwall.internal/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API token for user/service accounts/CLI.
        Use: `Authorization: Bearer <API_TOKEN>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-CTWALL_SESSION
      description: |
        Session JWT stored in HttpOnly cookie for GUI users.
    RefreshCookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-CTWALL_REFRESH
      description: |
        Refresh token stored in HttpOnly cookie for GUI session renewal.
  schemas:
...
    Error:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          maxLength: 120
          format: uri
          example: "about:blank"
        title:
          type: string
          maxLength: 120
          example: "Invalid Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          maxLength: 120
          example: "Field 'sbom_file' is missing."
        instance:
          type: string
          maxLength: 120
          example: "/api/v1/ingest"
...
    Product:
      type: object
      properties:
        id:
          type: string
          maxLength: 120
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 120
        createdAt:
          type: string
          maxLength: 120
          format: date-time
        updatedAt:
          type: string
          maxLength: 120
          format: date-time
        archivedAt:
          type: string
          maxLength: 120
          format: date-time
...
    AlertGroup:
      type: object
      required: [id, projectId, severity, category, type, status, groupKey, title, occurrences, firstSeenAt, lastSeenAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        severity: { type: string, enum: [INFO, WARN, ERROR] }
        category: { type: string }
        type: { type: string }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, CLOSED] }
        groupKey: { type: string }
        title: { type: string }
        entityRef: { type: string, nullable: true }
        occurrences: { type: integer, minimum: 0 }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        lastNotifiedAt: { type: string, format: date-time, nullable: true }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        acknowledgedBy: { type: string, format: uuid, nullable: true }
        closedAt: { type: string, format: date-time, nullable: true }
        closedBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /alert-groups/{id}:
    get:
      summary: Get alert group by id (project-scoped)
      operationId: getAlertGroup
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertGroup"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /products/{productId}:
    get:
      summary: Get Product details
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Delete Product
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Scopes ---
...